+++ app/src/main/java/com/smart/parts/picker/ui/assemblies/AssembliesFragment.kt
@@ -126,7 +126,7 @@ class AssembliesFragment : Fragment() {
         _binding = null
     }
 
-    private fun onAssemblyClick(assemblyId: Int) {
+    private fun onAssemblyClick(assemblyId: Long) {
         findNavController().navigate(
             AssembliesFragmentDirections.actionNavigationAssembliesToAssemblyBrowseFragment(
                 assemblyId
